Add the screen system and jump into it at startup.
This removes execution of the old Game1 class and replaces it with "Main"
which loads the ScreenManager as a component. ScreenManager should hold
the screens, including the GameplayScreen which will be modified to hold what
used to be Game1.

Unfortunately, the current structure of the components in Game1 is not fully
compatible with the architecture that ScreenManager and GameScreen impose
so modifying classes like Hero to be compatible with it is the first step.