Add GameObject wrappers for keyboard and mouse input.
This allows the objects to be automatically updated by iterating in
the GameplayScreen and is more extensible. For instance, if the game
becomes networked in the future, checking whether a remote player
performed an action does not check the keyboard and mouse on their
system, it queries data obtained from the network.