Add a GameObject base class for update and drawing.
This makes objects that get drawn or updated polymorphically the
same and they can be moved around and iterated over easily (rather
than having to explicitly use the Hero object, as well as each
projectile, tile, map, etc. They can all be treated the same when
appropriate.

The motiviation for this addition is primarly to allow GameObjects
to exist at the top-most level of the GameplayScreen. This will let
the GameplayScreen update and draw the objects it cares about.