Reorganize the hierarchy to be based on a GameObject
This puts the GameplayScreen more directly in control over drawing and updating objects in the game.
The GameplayScreen holds references to objects which need to be drawn or updated, while allowing the
objects to have owners one or two levels lower in the hierarchy. When they need to be added or removed,
the objects perform the action on the GameplayScreen inherited through the GameObject base class.

This commit also changes the way textures are loaded for objects lower in the hierarchy. Transient creations
like projectiles should not load their content every time they are spawned. Instead, their managing object (in
this case a subclass of Skill) should cache the necessary Textures and provide a reference to the appropriate
one during construction.